Split hilbert into header and source files
