+++ CMakeLists.txt
@@ -211,7 +211,7 @@ add_library(netket MODULE NetKet/pynetket.cc NetKet/Dynamics/py_dynamics.cpp
     NetKet/Graph/custom_graph.cc NetKet/Graph/abstract_graph.cc
     NetKet/Graph/py_graph.cc NetKet/Hilbert/hilbert_index.cc
     NetKet/Hilbert/bosons.cc NetKet/Hilbert/spins.cc NetKet/Hilbert/qubits.cc
-    NetKet/Hilbert/custom_hilbert.cc)
+    NetKet/Hilbert/custom_hilbert.cc NetKet/Hilbert/py_hilbert.cc)
 target_link_libraries(netket PUBLIC netket_lib pybind11)
 set_target_properties(netket PROPERTIES PREFIX "${PYTHON_MODULE_PREFIX}"
                                         SUFFIX "${PYTHON_MODULE_EXTENSION}")
